Rename ruby files to avoid smashing prawn files
Move all files aside from templates.rb to prawn/templates to
avoid smashing any existing files from prawn or pdf packages.
