+++ lib/prawn/templates/document/internals.rb
