+++ lib/prawn/templates/pdf/core/document_state.rb
