+++ lib/prawn/templates/pdf/core/errors.rb
