+++ lib/prawn/templates/pdf/core/object_store.rb
