+++ lib/prawn/templates/pdf/core/page.rb
