+++ lib/prawn/templates/text.rb
